+++ src/main/java/insane96mcp/experienceoverhaul/module/anvil/anvilrepair/AnvilBetterRepair.java
@@ -1,4 +1,4 @@
-package insane96mcp.experienceoverhaul.module.anvil;
+package insane96mcp.experienceoverhaul.module.anvil.anvilrepair;
 
 import insane96mcp.experienceoverhaul.module.EOModules;
 import insane96mcp.insanelib.core.feature.Feature;
@@ -7,11 +7,11 @@ import net.minecraft.world.item.ItemStack;
 
 import java.util.Optional;
 
-@LoadFeature(module = EOModules.ANVIL, description = "Allows custom anvil repair recipes defined via datapacks in data/<modid>/anvil_repairs/*.json.")
-public class AnvilCustomRepairs extends Feature {
+@LoadFeature(module = EOModules.ANVIL, description = "Makes repairing items cost less materials instead of an arbitrary 4. Allows custom anvil repair recipes defined via datapacks in data/<modid>/anvil_repairs/.")
+public class AnvilBetterRepair extends Feature {
 
     public static Optional<AnvilRepair> getCustomAnvilRepair(ItemStack left) {
-        if (!Feature.isEnabled(AnvilCustomRepairs.class))
+        if (!Feature.isEnabled(AnvilBetterRepair.class))
             return Optional.empty();
         for (AnvilRepair repair : AnvilRepairReloadListener.REPAIRS.values()) {
             if (repair.isItemToRepair(left))
@@ -21,7 +21,7 @@ public class AnvilCustomRepairs extends Feature {
     }
 
     public static Optional<AnvilRepair.RepairData> getCustomAnvilRepair(ItemStack left, ItemStack right) {
-        if (!Feature.isEnabled(AnvilCustomRepairs.class))
+        if (!Feature.isEnabled(AnvilBetterRepair.class))
             return Optional.empty();
         for (AnvilRepair repair : AnvilRepairReloadListener.REPAIRS.values()) {
             if (repair.isItemToRepair(left)) {